- Updated handling of file offset in linker (part 1)

+++ Source/Mosa.Compiler.Framework/BaseCallingConventionExtended.cs
@@ -101,7 +101,7 @@ namespace Mosa.Compiler.Framework
 
 				result += size;
 
-				Alignment.AlignUp(result, alignment);
+				Alignment.AlignUp(result, (uint)alignment);
 			}
 
 			return result;
